+++ test/docker/integration/nginx-redis.kong.conf
@@ -1,0 +1,4 @@
+set $session_storage             redis;
+set $session_redis_prefix        sessions;
+set $session_redis_host          kong-session-store;
+set $session_redis_port          6379;
