+++ test/docker/integration/setup.py
@@ -2,8 +2,10 @@
 
 import os
 import requests
-
 from collections import namedtuple
+
+from keycloak_client import KeycloakClient
+from kong_client import KongClient
 
 local_ip      = os.getenv("IP", default="")
 host          = "localhost"
@@ -11,6 +13,7 @@ env_file_path = ".env"
 
 Config = namedtuple("Config", [
     "keycloak_endpoint",
+    "keycloak_realm",
     "keycloak_admin",
     "keycloak_password",
     "client_id",
@@ -42,6 +45,7 @@ def get_config(env):
 
     return Config(
         keycloak_endpoint = keycloak_url,
+        keycloak_realm    = "master",
         keycloak_admin    = env["KEYCLOAK_USER"],
         keycloak_password = env["KEYCLOAK_PW"],
         client_id         = "kong",
@@ -52,107 +56,6 @@ def get_config(env):
         kong_endpoint     = "http://{}:{}".format(host, env["KONG_HTTP_ADMIN_PORT"])
     )
 
-
-class KeycloakClient:
-    def __init__(self, url, username, password):
-        self._endpoint = url
-        self._session  = requests.session()
-        self._username = username
-        self._password = password
-
-    def create_client(self, name, secret):
-        url     = "{}/auth/admin/realms/master/clients".format(self._endpoint)
-        payload = {
-            "clientId": name,
-            "secret": secret,
-            "redirectUris": ["*"],
-        }
-
-        headers = self.get_auth_header()
-        res     = self._session.post(url, json=payload, headers=headers)
-        
-        if res.status_code not in [201, 409]:
-            raise Exception("Cannot Keycloak create client")
-
-    def get_auth_header(self):
-        return {
-            "Authorization": "Bearer {}".format(self.get_admin_token())
-        }
-
-    def get_admin_token(self):
-        url     = "{}/auth/realms/master/protocol/openid-connect/token".format(self._endpoint)
-        
-        payload = "client_id=admin-cli&grant_type=password" + \
-            "&username={}&password={}".format(self._username, self._password)
-        
-        headers = {
-            "Content-Type": "application/x-www-form-urlencoded"
-        }
-        
-        res = self._session.post(url, data=payload, headers=headers)
-        res.raise_for_status()
-        
-        return res.json()["access_token"]
-
-
-class KongClient:
-    def __init__(self, url):
-        self._endpoint = url
-        self._session  = requests.session()
-
-    def create_service(self, name, upstream_url):
-        url = "{}/services".format(self._endpoint)
-        payload = {
-            "name": name,
-            "url": upstream_url,
-        }
-        res = self._session.post(url, json=payload)
-        res.raise_for_status()
-        return res.json()
-
-    def create_route(self, service_name, paths):
-        url = "{}/services/{}/routes".format(self._endpoint, service_name)
-        payload = {
-            "paths": paths,
-        }
-        res = self._session.post(url, json=payload)
-        res.raise_for_status()
-        return res.json()
-
-    def create_plugin(self, plugin_name, service_name, config):
-        url = "{}/services/{}/plugins".format(self._endpoint, service_name)
-        payload = {
-            "name": plugin_name,
-            "config": config,
-        }
-        res = self._session.post(url, json=payload)
-        try:
-            res.raise_for_status()
-        except Exception as e:
-            print(res.text)
-            raise e
-        return res.json()
-
-    def delete_service(self, name):
-        try:
-            routes = self.get_routes(name)
-            for route in routes:
-                self.delete_route(route)
-        except requests.exceptions.HTTPError:
-            pass
-        url = "{}/services/{}".format(self._endpoint, name)
-        self._session.delete(url).raise_for_status()
-
-    def delete_route(self, route_id):
-        url = "{}/routes/{}".format(self._endpoint, route_id)
-        self._session.delete(url).raise_for_status()
-
-    def get_routes(self, service_name):
-        url = "{}/services/{}/routes".format(self._endpoint, service_name)
-        res = self._session.get(url)
-        res.raise_for_status()
-        return map(lambda x: x['id'], res.json()['data'])
-
 if __name__ == '__main__':
     validate_ip_set()
 
@@ -162,6 +65,7 @@ if __name__ == '__main__':
 
     print("Creating Keycloak HTTP Client at {}".format(config.keycloak_endpoint))
     kc_client = KeycloakClient(config.keycloak_endpoint,
+                               config.keycloak_realm,
                                config.keycloak_admin,
                                config.keycloak_password)
